Extracted nuget.ts and removed q dependency
