+++ src/nerdbank-gitversioning.npm/package.json
@@ -22,7 +22,6 @@
     "gulp-tslint": "^5.0.0",
     "gulp-typescript": "2.13.6",
     "merge2": "1.0.2",
-    "q": "1.4.1",
     "tslint": "^3.11.0",
     "typescript": "1.8.10",
     "typings": "1.0.5"
