+++ src/nerdbank-gitversioning.npm/typings.json
@@ -1,7 +1,6 @@
 {
   "globalDependencies": {
-    "node": "registry:dt/node#6.0.0+20160608110640",
-    "q": "registry:dt/q#0.0.0+20160417152954"
+    "node": "registry:dt/node#6.0.0+20160608110640"
   },
   "dependencies": {
     "del": "registry:npm/del#2.2.0+20160416211512",
